ironing out generics
